Adding offset to UHI
