+++ boost_histogram/__init__.py
@@ -6,6 +6,6 @@ from ._hist import histogram
 
 from . import axis, storage, accumulators, algorithm, numpy
 
-from .utils import loc, rebin, project, underflow, overflow
+from .uhi import loc, rebin, project, underflow, overflow
 
 from .version import __version__
